# docs/swagger.yaml
basePath: /api
definitions:
  models.CreateListingRequest:
    properties:
      agent_id:
        example: agent_1001
        maxLength: 100
        minLength: 1
        type: string
      bathrooms:
        example: 3
        minimum: 0
        type: integer
      bedrooms:
        example: 3
        minimum: 0
        type: integer
      description:
        example: Modern penthouse with ocean view.
        maxLength: 2000
        type: string
      location:
        $ref: '#/definitions/models.Location'
      price:
        example: 185000000
        minimum: 0
        type: number
      title:
        example: Luxury 3-Bedroom Penthouse in Ikoyi
        maxLength: 255
        minLength: 3
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.ListingType'
        enum:
        - rent
        - sale
        - shortlet
        example: sale
    required:
    - agent_id
    - bedrooms
    - location
    - price
    - title
    - type
    type: object
  models.FieldError:
    properties:
      field:
        example: price
        type: string
      message:
        example: price must be greater than or equal to 0
        type: string
    type: object
  models.Listing:
    properties:
      agent_id:
        example: agent_1001
        type: string
      bathrooms:
        example: 3
        type: integer
      bedrooms:
        example: 3
        type: integer
      created_at:
        example: "2026-03-23T10:00:00Z"
        type: string
      description:
        example: Modern penthouse with ocean view and automated lighting.
        type: string
      distance_km:
        example: 2.45
        type: number
      id:
        example: 11111111-1111-1111-1111-111111111111
        type: string
      location:
        $ref: '#/definitions/models.Location'
      price:
        example: 185000000
        type: number
      title:
        example: Luxury 3-Bedroom Penthouse in Ikoyi
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.ListingType'
        example: sale
      updated_at:
        example: "2026-03-23T10:00:00Z"
        type: string
    type: object
  models.ListingType:
    enum:
    - rent
    - sale
    - shortlet
    type: string
    x-enum-varnames:
    - ListingTypeRent
    - ListingTypeSale
    - ListingTypeShortlet
  models.Location:
    properties:
      address:
        example: 12 Alexander Avenue, Ikoyi
        type: string
      city:
        example: Lagos
        type: string
      latitude:
        example: 6.452
        type: number
      longitude:
        example: 3.435
        type: number
    required:
    - latitude
    - longitude
    type: object
  models.PaginationMetadata:
    properties:
      current_page:
        example: 1
        type: integer
      has_next_page:
        example: true
        type: boolean
      has_prev_page:
        example: false
        type: boolean
      page_size:
        example: 10
        type: integer
      total_items:
        example: 42
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  models.StandardResponse:
    properties:
      data: {}
      error:
        example: ""
        type: string
      errors:
        items:
          $ref: '#/definitions/models.FieldError'
        type: array
      message:
        example: Operation successful
        type: string
      pagination:
        $ref: '#/definitions/models.PaginationMetadata'
      status_code:
        example: 200
        type: integer
      success:
        example: true
        type: boolean
    type: object
  models.UpdateListingRequest:
    properties:
      agent_id:
        example: agent_1001
        maxLength: 100
        minLength: 1
        type: string
      bathrooms:
        example: 3
        minimum: 0
        type: integer
      bedrooms:
        example: 3
        minimum: 0
        type: integer
      description:
        example: Updated description.
        maxLength: 2000
        type: string
      location:
        $ref: '#/definitions/models.UpdateLocationRequest'
      price:
        example: 175000000
        minimum: 0
        type: number
      title:
        example: Updated Penthouse Title
        maxLength: 255
        minLength: 3
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.ListingType'
        enum:
        - rent
        - sale
        - shortlet
        example: sale
    type: object
  models.UpdateLocationRequest:
    properties:
      address:
        example: 14 Alexander Avenue, Ikoyi
        type: string
      city:
        example: Lagos
        type: string
      latitude:
        example: 6.4525
        type: number
      longitude:
        example: 3.4355
        type: number
    type: object
info:
  contact: {}
  description: Production-ready Property Listings REST API with geospatial radius
    search, advanced filtering, and pagination.
  title: ExpertListing Property Management API
  version: "1.0"
paths:
  /health:
    get:
      description: Returns the service health status, uptime, and timestamp
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Service Health Check
      tags:
      - Health
  /listings:
    get:
      description: Returns paginated listings with optional sorting
      parameters:
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: created_at
        description: Sort field
        enum:
        - created_at
        - price
        - bedrooms
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Listing'
                  type: array
                pagination:
                  $ref: '#/definitions/models.PaginationMetadata'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: List property listings
      tags:
      - Listings
    post:
      consumes:
      - application/json
      description: Creates a new listing with nested location details, pricing, and
        agent metadata
      parameters:
      - description: Listing payload
        in: body
        name: listing
        required: true
        schema:
          $ref: '#/definitions/models.CreateListingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Listing'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/models.FieldError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Create a property listing
      tags:
      - Listings
  /listings/{id}:
    delete:
      description: Deletes a property listing by UUID
      parameters:
      - description: Listing UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Delete a listing
      tags:
      - Listings
    get:
      description: Returns the listing with matching UUID
      parameters:
      - description: Listing UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Listing'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Fetch a listing by ID
      tags:
      - Listings
    patch:
      consumes:
      - application/json
      description: Updates specific fields of an existing property listing
      parameters:
      - description: Listing UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Partial update payload
        in: body
        name: listing
        required: true
        schema:
          $ref: '#/definitions/models.UpdateListingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Listing'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/models.FieldError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Partially update a listing
      tags:
      - Listings
    put:
      consumes:
      - application/json
      description: Updates fields of an existing property listing
      parameters:
      - description: Listing UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Listing update payload
        in: body
        name: listing
        required: true
        schema:
          $ref: '#/definitions/models.UpdateListingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Listing'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/models.FieldError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Update a listing
      tags:
      - Listings
  /listings/search:
    get:
      description: Filters by type, price range, bedrooms, and returns listings within
        radius_km of a given coordinate
      parameters:
      - description: Listing type
        enum:
        - rent
        - sale
        - shortlet
        in: query
        name: type
        type: string
      - description: Minimum price
        in: query
        minimum: 0
        name: min_price
        type: number
      - description: Maximum price
        in: query
        minimum: 0
        name: max_price
        type: number
      - description: Exact bedroom count
        in: query
        minimum: 0
        name: bedrooms
        type: integer
      - description: Minimum bedroom count
        in: query
        minimum: 0
        name: min_bedrooms
        type: integer
      - description: City name
        in: query
        name: city
        type: string
      - description: Latitude for radius search (-90 to 90)
        in: query
        name: lat
        type: number
      - description: Longitude for radius search (-180 to 180)
        in: query
        name: lng
        type: number
      - description: Radius in kilometers
        in: query
        minimum: 0.1
        name: radius_km
        type: number
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: created_at
        description: Sort field
        enum:
        - created_at
        - price
        - bedrooms
        - distance
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Listing'
                  type: array
                pagination:
                  $ref: '#/definitions/models.PaginationMetadata'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/models.StandardResponse'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/models.FieldError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Search listings by filters and location radius
      tags:
      - Listings
swagger: "2.0"